+++ src/gui/common.rs
@@ -683,6 +683,7 @@ pub enum UndoSubject {
     CustomGameFile(usize, usize),
     CustomGameRegistry(usize, usize),
     CustomGameInstallDir(usize, usize),
+    CustomGameWinePrefix(usize, usize),
     BackupFilterIgnoredPath(usize),
     BackupFilterIgnoredRegistry(usize),
     RcloneExecutable,
@@ -711,6 +712,7 @@ impl UndoSubject {
             | UndoSubject::CustomGameFile(_, _)
             | UndoSubject::CustomGameRegistry(_, _)
             | UndoSubject::CustomGameInstallDir(_, _)
+            | UndoSubject::CustomGameWinePrefix(_, _)
             | UndoSubject::BackupFilterIgnoredPath(_)
             | UndoSubject::BackupFilterIgnoredRegistry(_)
             | UndoSubject::RcloneExecutable