+++ src/gui/shortcuts.rs
@@ -209,6 +209,7 @@ pub struct CustomGameHistory {
     pub files: Vec<TextHistory>,
     pub registry: Vec<TextHistory>,
     pub install_dir: Vec<TextHistory>,
+    pub wine_prefix: Vec<TextHistory>,
 }
 
 #[derive(Default)]
@@ -299,6 +300,7 @@ impl TextHistories {
             files: game.files.iter().map(|x| TextHistory::raw(x)).collect(),
             registry: game.registry.iter().map(|x| TextHistory::raw(x)).collect(),
             install_dir: game.install_dir.iter().map(|x| TextHistory::raw(x)).collect(),
+            wine_prefix: game.wine_prefix.iter().map(|x| TextHistory::raw(x)).collect(),
         };
         self.custom_games.push(history);
     }
@@ -347,6 +349,11 @@ impl TextHistories {
                 .custom_games
                 .get(*i)
                 .and_then(|x| x.install_dir.get(*j).map(|y| y.current()))
+                .unwrap_or_default(),
+            UndoSubject::CustomGameWinePrefix(i, j) => self
+                .custom_games
+                .get(*i)
+                .and_then(|x| x.wine_prefix.get(*j).map(|y| y.current()))
                 .unwrap_or_default(),
             UndoSubject::BackupFilterIgnoredPath(i) => self
                 .backup_filter_ignored_paths
@@ -414,6 +421,9 @@ impl TextHistories {
             UndoSubject::CustomGameInstallDir(i, j) => Box::new(Message::config(move |value| {
                 config::Event::CustomGameInstallDir(i, EditAction::Change(j, value))
             })),
+            UndoSubject::CustomGameWinePrefix(i, j) => Box::new(Message::config(move |value| {
+                config::Event::CustomGameWinePrefix(i, EditAction::Change(j, value))
+            })),
             UndoSubject::BackupFilterIgnoredPath(i) => Box::new(Message::config(move |value| {
                 config::Event::BackupFilterIgnoredPath(EditAction::Change(i, value))
             })),
@@ -455,6 +465,7 @@ impl TextHistories {
             UndoSubject::CustomGameFile(_, _) => "".to_string(),
             UndoSubject::CustomGameRegistry(_, _) => "".to_string(),
             UndoSubject::CustomGameInstallDir(_, _) => "".to_string(),
+            UndoSubject::CustomGameWinePrefix(_, _) => "".to_string(),
             UndoSubject::BackupFilterIgnoredPath(_) => "".to_string(),
             UndoSubject::BackupFilterIgnoredRegistry(_) => "".to_string(),
             UndoSubject::RcloneExecutable => TRANSLATOR.executable_label(),
@@ -474,6 +485,7 @@ impl TextHistories {
             | UndoSubject::RedirectTarget(_)
             | UndoSubject::CustomGameFile(_, _)
             | UndoSubject::CustomGameInstallDir(_, _)
+            | UndoSubject::CustomGameWinePrefix(_, _)
             | UndoSubject::BackupFilterIgnoredPath(_)
             | UndoSubject::RcloneExecutable => (!path_appears_valid(&current)).then_some(ERROR_ICON),
             UndoSubject::CustomGameName(_) | UndoSubject::CustomGameAlias(_) => {